chore: move market content out to separate component
